Review GLFW dependency + add phase1 explanation PDF

+++ src/algorithm/thistlethwaite/utilities/db_generator.cpp
@@ -3,7 +3,7 @@
 void DB_generator::generate(const Phase_info& phase_info, Database& database) const
 {
     Cube_bg_model cube;
-    uint8_t       depth = 0;
+    int           depth = 0;
     Timer         timer;
 
     timer.set();
@@ -12,7 +12,7 @@ void DB_generator::generate(const Phase_info& phase_info, Database& database) co
         // keeps track of visited nodes in each of the iterations
         std::vector<bool> visited(database.capacity(), false);
 
-        std::cout << "Depth: " << (int)depth << ". ";
+        std::cout << "Depth: " << depth << ". ";
 
         // this condition will be true only when database will become full
         if (_db_searcher(cube, Cube_bg_model::emove::NO_MOVE, phase_info, database, 0, depth, visited))
@@ -30,7 +30,7 @@ void DB_generator::generate(const Phase_info& phase_info, Database& database) co
 
 bool DB_generator::_db_searcher(Cube_bg_model cube, Cube_bg_model::emove last_move,
                                 const Phase_info& phase_info, Database& database,
-                                uint8_t depth, uint8_t max_depth,
+                                int depth, int max_depth,
                                 std::vector<bool>& visited) const
 {
     // id of the current state of cube
